+++ commons-core/src/main/java/com/wuzh/commons/core/reflect/ReflectUtils.java
@@ -16,7 +16,7 @@
 package com.wuzh.commons.core.reflect;
 
 import com.wuzh.commons.core.util.CalculationUtils;
-import com.wuzh.commons.core.util.JsonMapper;
+import com.wuzh.commons.core.json.jackson.JsonMapper;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;